backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
joblib==1.4.2
bcrypt==4.1.2
PyMySQL==1.1.0
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1

